+++ lib/index.js
@@ -437,7 +437,7 @@ function init ($module, $opts) {
 
   const filenames = [
     $module.filename,
-    require.resolve('./run-child.js'),
+    require.resolve('./runner-helpers/run-child.js'),
     require.resolve('../cli.js')
   ];
