+++ lib/runner-helpers/handle-fatal-message.js
@@ -3,6 +3,7 @@
 //npm
 const events = require('suman-events');
 const sumanUtils = require('suman-utils/utils');
+const colors = require('colors/safe');
 
 //project
 const resultBroadcaster = global.resultBroadcaster = (global.resultBroadcaster || new EE());
