+++ lib/runner-helpers/make-exit.js
@@ -4,20 +4,32 @@
 const events = require('suman-events');
 const sumanUtils = require('suman-utils/utils');
 const async = require('async');
+const _ = require('lodash');
+const AsciiTable = require('ascii-table');
 
 //project
+const constants = require('../../config/suman-constants');
 const debug = require('suman-debug')('s:runner');
 const resultBroadcaster = global.resultBroadcaster = (global.resultBroadcaster || new EE());
+const finalizeOutput = require('../helpers/finalize-output');
+
+
+function mapCopy (copy) {
+  return Object.keys(copy).map(key => {
+    const val = copy[key];
+    return val.value ? val.value : val.default;
+  });
+}
 
 //////////////////////////////////////////////////////////
 
 
-module.exports = function makeMakeExit(runnerObj) {
+module.exports = function makeMakeExit(runnerObj, tableRows) {
 
   return function makeExit (messages, timeDiff) {
 
-    debug('\n\n\n\tTable count:', tableCount);
-    debug('\tDone count:', doneCount);
+    debug('\n\n\n\tTable count:', runnerObj.tableCount);
+    debug('\tDone count:', runnerObj.doneCount);
 
     resultBroadcaster.emit(events.RUNNER_ENDED, new Date().toISOString());
 
@@ -49,7 +61,7 @@ module.exports = function makeMakeExit(runnerObj) {
 
     const totals = {
       SUMAN_IGNORE: '',
-      bailed: bailed ? 'YES' : 'no',
+      bailed: runnerObj.bailed ? 'YES' : 'no',
       suitesPassed: 0,
       suitesFailed: 0,
       testsPassed: 0,
