+++ lib/runner-helpers/on-exit.js
@@ -2,8 +2,10 @@
 
 //core
 const fs = require('fs');
+const path = require('path');
 
 //npm
+const colors = require('colors/safe');
 const events = require('suman-events');
 const sumanUtils = require('suman-utils/utils');
