+++ lib/runner-helpers/run-child.js
@@ -17,11 +17,11 @@ const EE = require('events');
 //npm
 const colors = require('colors/safe');
 const debug = require('suman-debug')('child');
+const sumanUtils = require('suman-utils/utils');
 
 //project
-const constants = require('../config/suman-constants');
-const sumanUtils = require('suman-utils/utils');
-const fatalRequestReply = require('./helpers/fatal-request-reply');
+const constants = require('../../config/suman-constants');
+const fatalRequestReply = require('../helpers/fatal-request-reply');
 
 //////////////////////////////////////////////////////////////////////////////
 
@@ -32,15 +32,6 @@ const projectRoot = global.projectRoot = process.env.SUMAN_PROJECT_ROOT;
 debug(' => child args => ', util.inspect(process.argv));
 debug(' => child env => ', util.inspect(process.env));
 
-
-// Object.defineProperty(global, 'resultBroadcaster', {
-//   __proto__: null, // no inherited properties
-//   set: function(v){
-//     console.log('v.a',v.a);
-//      console.error(new Error('setted').stack);
-//      this.value = v;
-//   }
-// });
 
 process.send = process.send || function (data) {
     console.error(colors.magenta('=> Suman implementation warning => Runner cannot receive data because process.send was not defined ' +
@@ -106,7 +97,7 @@ assert(sumanHelperDirRoot,
 
 //////////////////////////////////////////////////
 
-require('./helpers/log-stdio-of-child')(filePath);
+require('../helpers/log-stdio-of-child')(filePath);
 
 //////////////////////////////////////////////////////////
 
@@ -165,7 +156,7 @@ catch (err) {
 
 if (singleProc) {
   d.run(function () {
-    require('./handle-single-proc')(JSON.parse(process.env.SUMAN_SINGLE_PROCESS_FILES));
+    require('../handle-single-proc')(JSON.parse(process.env.SUMAN_SINGLE_PROCESS_FILES));
   });
 }
 else {
