+++ lib/runner.js
@@ -47,7 +47,6 @@ const os = require('os');
 const domain = require('domain');
 
 //npm
-const AsciiTable = require('ascii-table');
 const async = require('async');
 const _ = require('lodash');
 const readline = require('readline');
@@ -58,10 +57,7 @@ const events = require('suman-events');
 const debug = require('suman-debug')('s:runner');
 
 //project
-const handleTap = require('./runner-helpers/handle-tap');
-const noFilesFoundError = require('./helpers/no-files-found-error');
 const constants = require('../config/suman-constants');
-const finalizeOutput = require('./helpers/finalize-output');
 const ascii = require('./helpers/ascii');
 const sumanUtils = require('suman-utils/utils');
 const makeHandleBlocking = require('./runner-helpers/make-handle-blocking');
@@ -69,23 +65,30 @@ const resultBroadcaster = global.resultBroadcaster = (global.resultBroadcaster |
 const handleFatalMessage = require('./runner-helpers/handle-fatal-message');
 const logTestResult = require('./runner-helpers/log-test-result');
 const onExit = require('./runner-helpers/on-exit');
+const makeMakeExit = require('./runner-helpers/make-exit');
+const makeHandleIntegrantInfo = require('./runner-helpers/handle-integrant-info');
+const makeBeforeExit = require('./runner-helpers/make-before-exit-once-post');
+const makeSingleProcess = require('./runner-helpers/handle-single-process');
+const makeHandleMultipleProcesses = require('./runner-helpers/handle-multiple-processes');
 
 //////////////////////////////////////////////
 
 const _suman = global._suman;
 const cwd = process.cwd();
 const projectRoot = global.projectRoot = global.projectRoot || sumanUtils.findProjectRoot(cwd);
-const maxProcs = _suman.maxProcs;
+
 
 //TODO: https://github.com/mochajs/mocha/wiki/Third-party-reporters
 
 const messages = [];
-// const tableRows = [];
-const tableRows = {};
 const integrantHash = {};
 const integrantHashKeyValsForSumanOncePost = {};
 const userData = {}; // user will send data to runner for any/all tests, once before they exit
 const config = global.sumanConfig;
+const oncePosts = {};
+const allOncePostKeys = [];
+const  tableRows = {};
+const forkedCPs = [];
 
 const runnerObj = {
   doneCount:0,
@@ -95,19 +98,23 @@ const runnerObj = {
   startTime:null,
   endTime:null,
   bailed: false,
-  queuedCPs: []
+  queuedCPs: [],
+  hasOncePostFile: false,
+  innited: false,
+  oncePostModule:null,
+  oncePostModuleRet:null,
+  handleBlocking:null,
+  depContainerObj: null
 };
 
-var startTime,
-  endTime,
-  forkedCPs = [],
-  handleBlocking,
-  depContainerObj = null,
-  oncePostModule = null;
 
-var runnerObj = {
-  queuedCPs: []
-};
+
+const handleIntegrantInfo =
+  makeHandleIntegrantInfo(runnerObj, allOncePostKeys, integrantHash, integrantHashKeyValsForSumanOncePost);
+const makeExit =
+  makeMakeExit(runnerObj, tableRows);
+const beforeExitRunOncePost =
+  makeBeforeExit(runnerObj, oncePosts, integrantHashKeyValsForSumanOncePost, allOncePostKeys, userData);
 
 process.once('exit', onExit);
 
@@ -128,13 +135,11 @@ process.on('message', function (data) {
 
 
 function handleTableData (n, data) {
-
-  tableCount++;
+  runnerObj.tableCount++;
   tableRows[n.shortTestPath].tableData = data;
   n.send({
     info: 'table-data-received'
   });
-
 }
 
 function logTestData (data) {
@@ -161,275 +166,8 @@ function logTestData (data) {
 
 }
 
-function mapCopy (copy) {
-  return Object.keys(copy).map(key => {
-    const val = copy[key];
-    return val.value ? val.value : val.default;
-  });
-}
-
-
-
-const oncePosts = {};
-var oncePostModuleRet = null;
-var hasOncePostFile = false;
-const allOncePostKeys = [];
-var innited = false;
-
-function handleIntegrantInfo (msg, n) {
-
-  const oncePostKeys = msg.oncePost;
-
-  if (Number.isInteger(msg.expectedExitCode)) {
-    n.expectedExitCode = msg.expectedExitCode;
-  }
-  else if (msg.expectedExitCode !== undefined) {
-    throw new Error(' => Suman implementation error => expected exit code not an integer ' +
-      'for child process => ' + n.testPath);
-  }
-
-  if (Number.isInteger(msg.expectedTimeout)) {
-    if (!weAreDebugging) {
-      clearTimeout(n.to);
-      n.to = setTimeout(function () {
-        n.kill();
-      }, msg.expectedTimeout);
-    }
-
-  }
-  else if (msg.expectedTimeout !== undefined) {
-    throw new Error(' => Suman implementation error => expected timeout not an acceptable integer ' +
-      'for child process => ' + n.testPath);
-  }
-
-  //we want send back onlyPosts immediately because if we wait it blocks unnecessarily
-
-  assert(Array.isArray(oncePostKeys), 'oncePostKeys is not an array type.');
-  allOncePostKeys.push(oncePostKeys);
-
-  if (process.env.SUMAN_DEBUG === 'yes') {
-    console.log('\n', ' => Recevied integrant info msg =>', util.inspect(msg), '\nfrom testPath => ', n.testPath, '\n');
-  }
-
-  process.nextTick(function () {
-    n.send({
-      info: 'once-post-received'
-    });
-  });
-
-  if (oncePostKeys.length > 0 && !innited) {
-    try {
-      innited = true; //we only want to run this logic once
-      oncePostModule = require(path.resolve(global.sumanHelperDirRoot + '/suman.once.post.js'));
-      assert(typeof  oncePostModule === 'function', 'suman.once.post.js module does not export a function.');
-      hasOncePostFile = true;
-    }
-    catch (err) {
-      console.error(colors.red(' => Suman usage warning => you have suman.once.post data defined, ' +
-          'but no suman.once.post.js file.') + '\n' + (err.stack || err));
-    }
-
-  }
-
-  const integrants = msg.msg;
-
-  integrants.forEach(function (intg) {
-
-    if (!(String(intg) in integrantHash)) {
-      if (process.env.SUMAN_DEBUG === 'yes') {
-        console.log(' => intg with key ="' + String(intg) + '" is not in integrantHash => ', Object.keys(integrantHash));
-      }
-      integrantHash[String(intg)] = [];
-      integrantHash[String(intg)].push(n);   //store cps in hash, with integrant names as keys
-      verifyIntegrant(intg);
-    }
-    else if (String(integrantHash[intg]).toUpperCase() === 'READY') {
-
-      if (process.env.SUMAN_DEBUG === 'yes') {
-        console.log(' => integrants READY =>', util.inspect(msg));
-      }
-
-      n.send({info: 'integrant-ready', data: intg, val: integrantHashKeyValsForSumanOncePost[intg]});
-    }
-    else if (integrantHash[intg] instanceof Error) {
-
-      if (process.env.SUMAN_DEBUG === 'yes') {
-        console.log(' => integrants error =>', util.inspect(integrantHash[intg]));
-      }
-
-      n.send({info: 'integrant-error', data: integrantHash[intg].stack});
-    }
-    else if (Array.isArray(integrantHash[intg])) {
-
-      if (process.env.SUMAN_DEBUG === 'yes') {
-        console.log('\n', ' => child process with filePath =>', n.testPath, '\n is being push to integrants array value for key =>', intg);
-      }
-
-      integrantHash[intg].push(n);
-
-      if (process.env.SUMAN_DEBUG === 'yes') {
-
-        console.log('\n', 'integrantHash for key = "' + intg + '", looks like => \n',
-          integrantHash[intg].map(cp => 'cp with testPath =>' + cp.testPath))
-      }
-    }
-    else {
-      throw new Error('Unknown state of integrant readiness for integrant key => "' + intg + '",\n\n => ' + util.inspect(integrantHash));
-    }
-
-  });
-
-}
-
-const callbackOrPromise = require('./callback-or-promise');
-
-function beforeExitRunOncePost (cb) {
-
-  if (!hasOncePostFile) {
-    return process.nextTick(cb);
-  }
-
-  const flattenedAllOncePostKeys = _.uniq(_.flatten(allOncePostKeys));
-
-  if (sumanUtils.isSumanDebug()) {
-    console.error('integrantHashKeyValsForSumanOncePost =>', util.inspect(integrantHashKeyValsForSumanOncePost));
-  }
-  userData['suman.once.pre.js'] = integrantHashKeyValsForSumanOncePost;
-  const oncePostModuleRet = oncePostModule.apply(null, [userData]);
-
-  flattenedAllOncePostKeys.forEach(function (k) {
-    //we store an integer for analysis/program verification, but only really need to store a boolean
-    //for existing keys we increment by one, otherwise assign to 1
-    oncePosts[k] = oncePosts[k] || oncePostModuleRet[k];
-
-    if (typeof oncePosts[k] !== 'function') {
-
-      console.log(' => Suman is about to conk out =>\n\n' +
-        ' => here is the contents return by the exported function in suman.once.post.js =>\n\n', oncePosts);
-
-      throw new Error('\n' + colors.red(' => Suman usage warning => your suman.once.post.js ' +
-          'has keys whose values are not functions,\n\nthis applies to key ="' + k + '"'));
-
-    }
-  });
-
-  console.log('\n\n\n', ' => Suman message => All child processes exited.\n');
-
-  const keys = Object.keys(oncePosts);
-  if (keys.length) {
-    console.log('\n', ' => Suman message => Suman is now running the desired hooks in suman.once.post.js, which include => \n\t', colors.cyan(util.inspect(keys)));
-  }
-
-  async.eachSeries(keys, function (k, cb) {
-
-    callbackOrPromise(k, oncePosts, cb);
-
-  }, function (err) {
-    if (err) {
-      console.error(err.stack || err);
-      cb(err);
-    }
-    else {
-      console.log('\n\n', ' => Suman message => all suman.once.post.js hooks completed successfully...exiting...\n\n', '\n\n');
-      process.nextTick(function () {
-        cb(null);
-      });
-    }
-  });
-
-}
-
-
-
-function verifyIntegrant (intg) {
-
-  const d = domain.create();
-
-  d.once('error', function (err) {
-
-    console.log(err.stack || err);
-    const cps = integrantHash[intg];
-
-    integrantHash[intg] = err;
-
-    console.log(!Array.isArray(cps) ? '\n => for intg ="' + intg + '", cps in error => ' + util.inspect(cps) : '');
-
-    cps.forEach(function (cp) {
-      // TODO cps does not seem to be an array here
-      cp.send({info: 'integrant-error', data: err});
-    });
-
-    throw new Error(' => Suman usage error in suman.once.js for key => "' + intg + '"\n' + (err.stack || err));
-  });
-
-  var callable = true;
-
-  function sendOutMsg (val) {
-
-    if (!callable) {
-      //make sure this function is only called once per intg key
-      console.error(' => Suman usage warning in your suman.once.js file => \n' +
-        '=> callback was fired twice for key = "' + intg + '"');
-      return;
-    }
-
-    callable = false;
-    const cps = integrantHash[intg];
-
-    if (process.env.SUMAN_DEBUG === 'yes') {
-      console.log(!Array.isArray(cps) ? '\n => for intg ="' + intg + '", cps => ' + util.inspect(cps) : '');
-    }
-
-    integrantHash[intg] = 'READY';
-    integrantHashKeyValsForSumanOncePost[intg] = val;
-
-    if (sumanUtils.isSumanDebug()) {
-      console.log(' => sending out READY message for integrant = "' + intg + '" to the following cps => \n', cps.map(function (cp) {
-        return cp.testPath;
-      }), '\n\n');
-    }
-
-    cps.forEach(function (cp) {
-      cp.send({info: 'integrant-ready', data: intg, val: val});
-    });
-
-  }
-
-  d.run(function () {
-    process.nextTick(function () {
-      const fn = depContainerObj[intg];
-      assert(typeof fn === 'function', 'Integrant listing is not a function => ' + intg);
-      if (fn.length > 0) {
-        fn.apply(global, [function (err, val) {
-          if (err) {
-            //TODO: fix this, need to handle error properly
-            console.error(err.stack || err);
-            d.emit('error', err);
-          }
-          else {
-            // TODO: assert that value has been serialized (string, number, boolean, etc)
-            sumanUtils.runAssertionToCheckForSerialization(val);
-            sendOutMsg(val);
-          }
-        }]);
-      }
-      else {
-        Promise.resolve(fn.apply(global, [])).then(function (val) {
-          // TODO: assert that value has been serialized (string, number, boolean, etc)
-          sumanUtils.runAssertionToCheckForSerialization(val);
-          sendOutMsg(val);
-        }, function (err) {
-          console.error(err.stack || err);
-          d.emit('error', err);
-        });
-      }
-    });
-  });
-}
 
 function handleMessageForSingleProcess (msg, n) {
-
-  if (listening) {
 
     switch (msg.type) {
 
@@ -470,16 +208,9 @@ function handleMessageForSingleProcess (msg, n) {
       default:
         throw new Error(' => Suman internal error => bad msg.type in runner');
     }
-
-  }
-  else {
-    throw new Error('Suman internal error => this definitely shouldn\'t happen');
-  }
 }
 
 function handleMessage (msg, n) {
-
-  if (listening) {
 
     switch (msg.type) {
 
@@ -520,413 +251,16 @@ function handleMessage (msg, n) {
         throw new Error(' => Suman implementation error => Bad msg.type in runner, perhaps the user sent a message with process.send?');
     }
 
-  }
-  else {
-    throw new Error(' => Suman implementation error => this definitely shouldn\'t happen, please report.');
-  }
 }
 
-function runAllTestsInSingleProcess (runObj) {
 
-  const args = [];
+const runSingleOrMultipleDirs =
+  makeHandleMultipleProcesses(runnerObj, tableRows, messages, forkedCPs, handleMessage, beforeExitRunOncePost, makeExit);
 
-  //handle if dirs is not an array
-  // var files = getFilePaths(_.flattenDeep([dirs]));
+const runAllTestsInSingleProcess =
+  makeSingleProcess(runnerObj, handleMessageForSingleProcess, messages, beforeExitRunOncePost, makeExit);
 
-  // files = _.flattenDeep([files]);
 
-  var files = runObj.files;
-
-  if (global.sumanOpts.rand) {
-    files = _.shuffle(files);
-  }
-
-  const $files = sumanUtils.removeSharedRootPath(files);
-  const SUMAN_SINGLE_PROCESS_FILES = JSON.stringify($files);
-
-  const toPrint = $files.map(function (f) {
-    return ' => ' + f[1];
-  });
-
-  toPrint.unshift('');
-  toPrint.push('');
-  toPrint.push('');
-  toPrint.push('');  // add some vertical padding
-
-  console.log(' => Suman files running in single process =>\n', toPrint.join('\n\t'));
-  startTime = Date.now();
-
-  const sumanEnv = Object.assign({}, process.env, {
-    SUMAN_CONFIG: JSON.stringify(global.sumanConfig),
-    SUMAN_OPTS: JSON.stringify(global.sumanOpts),
-    SUMAN_SINGLE_PROCESS_FILES: SUMAN_SINGLE_PROCESS_FILES,
-    SUMAN_SINGLE_PROCESS: 'yes',
-    SUMAN_RUNNER: 'yes',
-    SUMAN_RUN_ID: _suman.runId,
-    SUMAN_RUNNER_TIMESTAMP: _suman.timestamp,
-    NPM_COLORS: process.env.NPM_COLORS || (global.sumanOpts.no_colors ? 'no' : 'yes')
-  });
-
-  if (global.sumanOpts.register) {
-    args.push('--register');
-  }
-
-  const execArgz = ['--expose-gc', '--harmony', '--expose_debug_as=v8debug'];
-
-  if (weAreDebugging) {
-    if (!global.sumanOpts.ignore_break) {  //NOTE: this allows us to focus on debugging runner
-      execArgz.push('--debug-brk');
-    }
-    execArgz.push('--debug=' + (5303 + processId++));
-  }
-
-  const ext = _.merge({}, {
-    cwd: projectRoot,  //TODO: improve this logic
-    silent: !(global.sumanOpts.no_silent === true),
-    execArgv: execArgz,
-    env: sumanEnv,
-    // uid: gid++,
-    detached: false   //TODO: detached:false works but not true
-  });
-
-  const n = cp.fork(path.resolve(__dirname + '/run-child.js'), args, ext);
-
-  //TODO: n.testPath is not defined, have to mititage this so that logic still works
-
-  n.on('message', function (msg) {
-    handleMessageForSingleProcess(msg, n);
-  });
-
-  n.on('error', function (err) {
-    throw new Error(err.stack);
-  });
-
-  if (global.sumanOpts.no_silent !== true) {
-
-    n.stdio[2].setEncoding('utf-8');
-    n.stdio[2].on('data', function (data) {
-
-      const d = String(data).split('\n').map(function (line) {
-        return '[' + '???' + '] ' + line;
-      }).join('\n');
-
-      global.sumanStderrStream.write('\n\n');
-      global.sumanStderrStream.write(d);
-
-      if (weAreDebugging) {  //TODO: add check for NODE_ENV=dev_local_debug
-        //TODO: go through code and make sure that no console.log statements should in fact be console.error
-        console.log('pid => ', n.pid, 'stderr => ', d);
-      }
-
-    });
-
-  }
-
-  n.on('exit', function (code, signal) {
-
-    if (process.env.SUMAN_DEBUG === 'yes') {
-      console.log('\n', colors.black.bgYellow(' => process given by => ' + n.shortTestPath + ' exited with code: ' + code + ' '), '\n');
-    }
-
-    if (process.env.SUMAN_DEBUG === 'yes') {
-      global.timeOfMostRecentExit = Date.now();
-    }
-
-    n.removeAllListeners();
-
-    doneCount++;
-    messages.push({code: code, signal: signal});
-    // tableRows[n.shortTestPath].actualExitCode = code;
-
-    //TODO: if bail, need to make that clear to user here
-
-    listening = false;
-    setImmediate(function () {
-      beforeExitRunOncePost(function (err) {
-        makeExit(messages, Date.now() - startTime);
-      });
-    });
-
-  });
-
-}
-
-function runSingleOrMultipleDirs (runObj) {
-
-  const args = [];
-
-  if (global.usingLiveSumanServer) {
-    args.push('--live_suman_server');
-  }
-
-  var files = runObj.files;
-  const filesThatDidNotMatch = runObj.filesThatDidNotMatch;
-
-  filesThatDidNotMatch.forEach(function (val) {
-    console.log('\n', colors.bgBlack.yellow(' => Suman message =>  A file in a relevant directory ' +
-      'did not match your regular expressions => '), '\n', util.inspect(val));
-  });
-
-  //TODO: need to remove duplicate files before calling resultBroadcaster
-  resultBroadcaster.emit(events.RUNNER_STARTED, files.length);
-
-  if (global.sumanOpts.rand) {
-    files = _.shuffle(files);
-  }
-
-  //TODO: need to make sure list of files is unique list, if not report that as non-fatal error
-
-  handleBlocking.determineInitialStarters(files);
-  startTime = Date.now();
-
-  const fileObjArray = sumanUtils.removeSharedRootPath(files);
-
-  const sumanEnv = Object.assign({}, process.env, {
-    SUMAN_CONFIG: JSON.stringify(global.sumanConfig),
-    SUMAN_OPTS: JSON.stringify(global.sumanOpts),
-    SUMAN_RUNNER: 'yes',
-    SUMAN_RUN_ID: _suman.runId,
-    SUMAN_RUNNER_TIMESTAMP: _suman.timestamp,
-    NPM_COLORS: process.env.NPM_COLORS || (global.sumanOpts.no_color ? 'no' : 'yes')
-  });
-
-  const execFile = path.resolve(__dirname + '/run-child.js');
-
-  fileObjArray.forEach(function (fileShortAndFull, index) {
-
-    const file = fileShortAndFull[0];
-    const shortFile = fileShortAndFull[1];
-
-    // const basename = path.basename(file);
-
-    var basename = file.length > 28 ? ' ' + String(file).substring(Math.max(0, file.length - 28)) + ' ' : file;
-
-    const m = String(basename).match(/\//g);
-
-    if (m && m.length > 1) {
-      const arr = String(basename).split('');
-      var i = 0;
-      while (arr[i] !== '/') {
-        arr.shift();
-      }
-      basename = arr.join('');
-    }
-
-    tableRows[shortFile] = {
-      actualExitCode: null,
-      shortFilePath: shortFile,
-      tableData: null,
-      defaultTableData: {
-        SUITES_DESIGNATOR: basename
-      }
-    };
-
-    const argz = JSON.parse(JSON.stringify(args));
-
-    function run () {
-
-      const execArgz = ['--expose-gc', '--harmony'];
-
-      if (weAreDebugging) {
-        if (!global.sumanOpts.ignore_break) {  //NOTE: this allows us to focus on debugging runner
-          execArgz.push('--debug-brk');
-        }
-        execArgz.push('--debug=' + (5303 + processId++));
-      }
-
-      const ext = _.merge({env: {SUMAN_CHILD_TEST_PATH: file}}, {
-        cwd: global.sumanOpts.force_cwd_to_be_project_root ? projectRoot : path.dirname(file),  //TODO: improve this logic
-        // silent: !(global.sumanOpts.no_silent === true),
-        // silent: false,
-        // stdio: ['ignore', 'ignore', 'ignore', 'ipc'],
-        stdio: ['ignore', 'pipe', 'pipe', 'ipc'],
-        execArgv: execArgz,
-        env: sumanEnv,
-        detached: false   //TODO: detached:false works but not true
-      });
-
-      var n;
-
-      const extname = path.extname(shortFile);
-
-      if (global.sumanOpts.library_coverage) {
-        console.log(' => Running library coverage with Istanbul for index => ' + index + ' file => ' + shortFile);
-        const coverageDir = path.resolve(global.projectRoot + '/coverage/runner+' + String(shortFile).replace(/\//g, '-'));
-        n = cp.spawn('istanbul', ['cover', execFile, '--dir', coverageDir, '--'].concat(args), ext);
-      }
-      else if ('.js' === extname) {
-
-        // n = cp.fork(execFile, argz, ext);
-        argz.unshift(execFile);
-        n = cp.spawn('node', argz, ext);
-      }
-      // else if('.sh'  === extname){
-      //   argz.unshift(file);
-      //   console.log('args => ', argz);
-      //   n = cp.spawn('exec', argz, ext);
-      // }
-      else if ('.sh' === extname) {
-        // we run the file directly, hopefully it has a hashbang
-        n = cp.spawn(file, argz, ext);
-      }
-      else if ('.bash' === extname) {
-        argz.unshift(file);
-        n = cp.spawn('bash', argz, ext);
-      }
-      else {
-        throw new Error(' => File with that extension is not yet supported by the Suman test runner => ' + shortFile);
-      }
-
-      if (!weAreDebugging) {
-        n.to = setTimeout(function () {
-          n.kill('SIGKILL');
-        }, 60000);
-      }
-
-      n.testPath = file;
-      n.shortTestPath = shortFile;
-
-      forkedCPs.push(n);
-
-      n.on('message', function (msg) {
-        handleMessage(msg, n);
-      });
-
-      n.on('error', function (err) {
-        console.error('\n', err.stack || err, '\n');
-      });
-
-      if (n.stdio /*false && global.sumanOpts.no_silent !== true*/) {
-
-        n.stdout.setEncoding('utf8');
-        n.stderr.setEncoding('utf8');
-
-        if (false) {
-          n.stdout.pipe(process.stdout);
-          n.stderr.pipe(process.stderr);
-        }
-
-        if (true || global.useTAPOutput) {
-          n.stdout.pipe(handleTap());
-        }
-
-        n.stdio[2].setEncoding('utf-8');
-        n.stdio[2].on('data', function (data) {
-
-          const d = String(data).split('\n').filter(function (line) {
-            return String(line).length;
-          }).map(function (line) {
-            return '[' + n.shortTestPath + '] ' + line;
-          }).join('\n');
-
-          global.sumanStderrStream.write('\n\n');
-          global.sumanStderrStream.write(d);
-
-          if (weAreDebugging) {  //TODO: add check for NODE_ENV=dev_local_debug
-            //TODO: go through code and make sure that no console.log statements should in fact be console.error
-            console.log('pid => ', n.pid, 'stderr => ', d);
-          }
-
-        });
-
-      }
-      else {
-        console.log(' => Stdio not available for child process.');
-      }
-
-      n.once('exit', function (code, signal) {
-
-        resultBroadcaster.emit(events.TEST_FILE_CHILD_PROCESS_EXITED, {
-          testPath: n.testPath,
-          exitCode: code
-        });
-
-        if (process.env.SUMAN_DEBUG === 'yes') {
-          console.log('\n', colors.black.bgYellow(' => process given by => ' + n.shortTestPath + ' exited with code: ' + code + ' '), '\n');
-        }
-
-        if (process.env.SUMAN_DEBUG === 'yes') {
-          global.timeOfMostRecentExit = Date.now();
-        }
-
-        n.removeAllListeners();
-
-        const originalExitCode = JSON.parse(JSON.stringify(code));
-
-        if (n.expectedExitCode !== undefined) {
-          if (code === n.expectedExitCode) {
-            code = 0;
-          }
-          // else{  // do not need this because child process should handle this
-          //    code = constants.EXIT_CODES.EXPECTED_EXIT_CODE_NOT_MET;
-          // }
-        }
-
-        doneCount++;
-        messages.push({code: code, signal: signal});
-        tableRows[n.shortTestPath].actualExitCode = n.expectedExitCode !== undefined ?
-          n.expectedExitCode + '/' + originalExitCode : originalExitCode;
-
-        //TODO: if bail, need to make that clear to user here
-        if ((bailed = (code > 0 && global.sumanOpts.bail)) || (doneCount >= forkedCPs.length && runnerObj.queuedCPs.length < 1)) {
-          endTime = Date.now();
-          listening = false;
-          setImmediate(function () {
-            beforeExitRunOncePost(function (err) {
-              makeExit(messages, endTime - startTime);
-            });
-          });
-        }
-        else {
-          const testPath = n.testPath;
-          handleBlocking.releaseNextTests(testPath, runnerObj);
-          if (process.env.SUMAN_DEBUG === 'yes') {
-            console.log(' => Time required to release next test(s) => ', Date.now() - global.timeOfMostRecentExit, 'ms');
-          }
-        }
-      });
-
-    }
-
-    run.testPath = file;
-    run.shortTestPath = shortFile;
-
-    if (handleBlocking.shouldFileBeBlockedAtStart(file)) {
-      runnerObj.queuedCPs.push(run);
-      // argz.push('--blocked');
-      if (process.env.SUMAN_DEBUG == 'on') {
-        console.log('File is blocked =>', file);
-      }
-    }
-    else {
-      run();
-      if (process.env.SUMAN_DEBUG == 'on') {
-        console.log(' => File is running =>', file);
-      }
-    }
-
-  });
-
-  if (forkedCPs.length < 1 && runnerObj.queuedCPs.length > 0) {
-    throw new Error(' => Suman internal error => fatal start order algorithm error, please file an issue on Github, thanks.');
-  }
-
-  if (forkedCPs.length < 1) {
-    noFilesFoundError(files);
-  }
-  else {
-    const totalCount = forkedCPs.length + runnerObj.queuedCPs.length;
-    var suites = totalCount === 1 ? 'suite' : 'suites';
-    var processes = totalCount === 1 ? 'process' : 'processes';
-    resultBroadcaster.emit(events.RUNNER_INITIAL_SET, forkedCPs, processes, suites);
-    const addendum = maxProcs < totalCount ? ' with no more than ' + maxProcs + ' running at a time.' : '';
-    resultBroadcaster.emit(events.RUNNER_OVERALL_SET, totalCount, processes, suites, addendum);
-  }
-
-  if (global.sumanOpts.errors_only) {
-    resultBroadcaster.emit(events.ERRORS_ONLY_OPTION);
-  }
-}
 
 function findTestsAndRunThem (runObj, runOnce, $order) {
 
@@ -935,14 +269,14 @@ function findTestsAndRunThem (runObj, runOnce, $order) {
   //need to get rid of this property so child processes cannot require Suman index file
   delete process.env.SUMAN_EXTRANEOUS_EXECUTABLE;
 
-  handleBlocking = makeHandleBlocking(_.mapValues($order, function (val) {
+  runnerObj.handleBlocking = makeHandleBlocking(_.mapValues($order, function (val) {
     val.testPath = path.resolve(projectRoot + '/' + val.testPath);
     return val;
   }));
 
   process.nextTick(function () {
 
-    depContainerObj = runOnce();  //TODO: should this be done  before this point in the program?
+    runnerObj.depContainerObj = runOnce();  //TODO: should this be done  before this point in the program?
     resultBroadcaster.emit(events.RUNNER_ASCII_LOGO, ascii.suman_runner);
 
     if (process.env.SUMAN_SINGLE_PROCESS === 'yes') {
